requirements.txt
# Core dependencies
strands-agents>=0.1.0
boto3>=1.34.0
bedrock-agentcore>=0.1.0
pydantic>=2.0.0
mcp>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0